For convenient development, we enrich initialization with code to add the inspector automatically.
We then add an "expanded?" input cell, and have the kids logic
check that to decide if it should add inspector nodes for each
child of the current node.

We then add an "on click" handler that toggles the "expanded?" cell
of the clicked node.

The trick there ^^ was first navigating to the top div in a node, and
two solutions are provided.

The other trick was knowing enough HTML to stopPropagation.